fixed Carrot Rifle not using ammo in survival mode

+++ GardenArsenalForgeTry/src/main/java/multiteam/gardenarsenal/setup/weapons/CarrotRifle.java
@@ -9,6 +9,8 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.AbstractArrowEntity;
 import net.minecraft.entity.projectile.SnowballEntity;
+import net.minecraft.item.ArrowItem;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.util.SoundCategory;
@@ -58,8 +60,9 @@ public class CarrotRifle extends ModWeapons {
         if (!(player instanceof PlayerEntity)) {
             return;
         }
+        boolean hasAmmo = ((PlayerEntity) player).inventory.hasItemStack(new ItemStack(Items.CARROT));
         World worldIn = player.world;
-        if (!worldIn.isRemote && timeR >= 5) {
+        if (!worldIn.isRemote && timeR >= 5 && hasAmmo) {
             SnowballEntity snowballentity = new SnowballEntity(worldIn, player);
             snowballentity.setItem(new ItemStack(Items.CARROT));
             snowballentity.func_234612_a_(player, player.rotationPitch, player.rotationYaw, 0.0F, 2.0F, 1.0F);
@@ -67,11 +70,14 @@ public class CarrotRifle extends ModWeapons {
             worldIn.playSound((PlayerEntity)null, player.getPosX(), player.getPosY(), player.getPosZ(), SoundEvents.BLOCK_ANCIENT_DEBRIS_HIT, SoundCategory.NEUTRAL, 0.5F, 0.4F / (random.nextFloat() * 0.4F + 0.8F));
             timeR = 0;
 
-            ItemStack ammoStack = new ItemStack(Items.CARROT);
+            ItemStack ammoStack = ((PlayerEntity) player).inventory.getStackInSlot(((PlayerEntity) player).inventory.getSlotFor(new ItemStack(Items.CARROT)));
 
-            ammoStack.shrink(1);
-            if (ammoStack.isEmpty()) {
-                ((PlayerEntity) player).inventory.deleteStack(ammoStack);
+            // use an ammo
+            if(!((PlayerEntity) player).isCreative()){
+                ammoStack.shrink(1);
+                if (ammoStack.isEmpty()) {
+                    ((PlayerEntity) player).inventory.deleteStack(ammoStack);
+                }
             }
         }
         timeR++;
